Now protected endpoints so user must be logged in for POST/PUT/DELETE of sightings

+++ sightings/views.py
@@ -2,12 +2,14 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.exceptions import NotFound
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from .models import Sighting
 from .serializers.common import SightingSerializer
 from .serializers.populated import PopulatedSightingSerializer
 
 class SightingListView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     def get(self, _request):
         sighting = Sighting.objects.all()
@@ -15,7 +17,9 @@ class SightingListView(APIView):
         return Response(serialized_sighting.data, status=status.HTTP_200_OK)
     
     def post(self, request):
+        request.data['owner'] = request.user.id
         sighting_to_add = SightingSerializer(data=request.data)
+
         try:
             sighting_to_add.is_valid()
             sighting_to_add.save()
@@ -25,6 +29,7 @@ class SightingListView(APIView):
             return Response(e.__dict__ if e.__dict__ else str(e), status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
 class SightingDetailView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     def get_sighting(self, pk):
         try:
@@ -41,6 +46,9 @@ class SightingDetailView(APIView):
         sighting_to_edit = self.get_sighting(pk=pk)
         updated_sighting = SightingSerializer(sighting_to_edit, data=request.data)
 
+        if sighting_to_edit.owner != request.user:
+            return Response({'message': 'Unauthorized action'}, status=status.HTTP_401_UNAUTHORIZED)
+
         if updated_sighting.is_valid():
             updated_sighting.save()
             return Response(updated_sighting.data, status=status.HTTP_202_ACCEPTED)
@@ -49,5 +57,9 @@ class SightingDetailView(APIView):
     
     def delete(self, request, pk):
         sighting_to_delete = self.get_sighting(pk=pk)
+
+        if sighting_to_delete.owner != request.user and not (request.user.is_staff or request.user.is_admin):
+            return Response({'message': 'Unauthorized action'}, status=status.HTTP_401_UNAUTHORIZED)
+
         sighting_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
